pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quarry"
version = "2.0.0"
description = "An intuitive web scraping toolkit with intelligent templates and interactive workflows"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "beautifulsoup4",
    "pandas",
    "pyarrow",
    "pyyaml",
    "typer",
    "questionary",
    "pydantic",
    "rich",
    "click",
]

[project.scripts]
quarry = "quarry.quarry:main"
"quarry.scout" = "quarry.tools.scout.cli:scout"
"quarry.survey" = "quarry.tools.survey.cli:survey"
"quarry.excavate" = "quarry.tools.excavate.cli:excavate"
"quarry.polish" = "quarry.tools.polish.cli:polish"
"quarry.ship" = "quarry.tools.ship.cli:ship"

[tool.setuptools]
packages = ["quarry"]

[tool.setuptools.package-data]
quarry = ["**/*.json", "**/*.yml", "**/*.yaml"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "preserve"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "PL", "RUF"]
ignore = []

# Relax linting for interactive CLI modules where pragmatic patterns are used
[tool.ruff.lint.per-file-ignores]
"quarry/tools/**" = [
    # Long rich-format strings and prompts
    "E501",
    # Import-outside-toplevel for lazy/circular-avoidance imports
    "PLC0415",
    # Function complexity for interactive flows
    "PLR0911", "PLR0912", "PLR0913", "PLR0915",
    # Magic values are common in CLI thresholds/prompts
    "PLR2004",
    # Overwriting loop variables in simple transformations
    "PLW2901",
    # Unused noqa markers from earlier iterations
    "RUF100",
]

"quarry/framework_profiles/**" = [
    # Heuristic thresholds and pattern counts are intentional
    "PLR2004",
    # Large, branching detection logic
    "PLR0911", "PLR0912", "PLR0915",
    # Long lines for lambda-heavy selectors and comments
    "E501",
]

"quarry/wizard.py" = ["PLR0912", "PLR0915"]
"quarry/quarry.py" = ["PLR0913"]
"quarry/inspector.py" = ["PLR0911", "PLR0912", "PLR2004"]
"quarry/lib/selectors.py" = ["PLR0911"]
"quarry/lib/prompts.py" = ["PLR0912"]
"quarry/lib/http.py" = ["PLR0912", "PLR0915", "PLR0913", "PLR2004"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual typing - will enable module by module
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false  # Disabled during gradual migration
strict_equality = true
plugins = ["pydantic.mypy"]
exclude = "^(tests/|scripts/|examples/)"

# Per-module options for gradual typing
[[tool.mypy.overrides]]
module = "quarry.framework_profiles.*"
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "quarry.http"
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Suppress typing noise in non-library code paths while keeping the library strict
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "scripts.*",
    "examples.*",
]
ignore_errors = true

# External dependencies without type stubs
[[tool.mypy.overrides]]
module = [
    "bs4.*",
    "questionary.*",
    "pyarrow.*",
    "pandas.*",
    "psycopg.*",
    "psycopg_pool.*",
]
ignore_missing_imports = true

